requirements.txt
pytest==7.4.0
Appium-Python-Client==2.11.1
pytest-html==3.2.0
requests==2.31.0
